requirements.txt
langgraph>=0.0.20
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.2
langchain-core>=0.1.0
pydantic>=2.6.1
python-dotenv>=1.0.0
openai>=1.12.0